Args for Redis uri and standard Conversation creds. Misc.
Use CONVERSATION_USERNAME, CONVERSATION_PASSWORD and WORKSTATION_ID
for the Conv. args since those are standard in the SDK.

Use REDIS_URI to connect to Compose for Redis (as documented).
Keep REDIS_PORT and REDIS_IP as fallback for local Redis.
Attempt Redis client w/ no params if none of the above.

Misc fixes for typo or unused vars.
Additional logging for trouble-shooting.