New workspace to avoid warnings and hidden nodes.
The original workspace had some nodes that were not accessible
from the UI and also generated a warning at runtime.  Something
was corrupt in there.

This is a rebuilt version. Same intents/entities. Recreated the
dialog.

Uses the beta @sys-location so that we are not limited to 4 cities in Germany.

Uses output.action.lookupWeather which should be closer to the sample for
how to pass an action back to the app (this is implemented in the
code in prior commit).